Did slight refactoring and added documentation.
