+++ src/graphic/MenuBar.java
@@ -4,12 +4,22 @@ import java.awt.Font;
 
 import javax.swing.*;
 
-public class OptionMenu extends JMenuBar{
+/**
+ * The Menu Bar for the program.
+ * @author nguyet15
+ *
+ */
+public class MenuBar extends JMenuBar{
 	private FileMenu file;
 	private EditMenu edit;
 	private ViewMenu view;
 	
-	public OptionMenu() {
+	/**
+	 * Constructor for the menu bar.
+	 * 
+	 * @effect: Create an instance of the menu bar.
+	 */
+	public MenuBar() {
 		file = new FileMenu();
 		add(file);
