+++ src/graphic/ViewMenu.java
@@ -2,11 +2,21 @@ package graphic;
 
 import javax.swing.*;
 
+/**
+ * Add the options related to view to the menu bar.
+ * @author nguyet15
+ *
+ */
 public class ViewMenu extends JMenu{
 	private JMenuItem zoomIn;
 	private JMenuItem zoomOut;
 	private JMenuItem reset;
 	
+	/**
+	 * Constructor for the class
+	 * 
+	 * @effect: Create the ViewMenu class to be added to the menu bar
+	 */
 	public ViewMenu() {
 		setText("View");
